+++ webpack-dev.config.js
@@ -36,6 +36,12 @@ module.exports = merge(common, {
             hash: true,
             inject: 'head',
         }),
+        new HtmlWebpackPlugin({
+            filename: 'isolation.html',
+            template: 'example/test-isolation.html',
+            hash: true,
+            inject: 'head',
+        }),
     ],
     devServer: {
         contentBase: path.resolve(__dirname, '_bundles'),